refactor(core): part directives to class feature
- changed `StylableDirectives` to `StPartDirectives` and moved to css class feature
- remove `features/types.ts`
- finally removed `stylable-value-parsers`
