--- packages/core/src/features/types.ts
@@ -1,12 +1,0 @@
-import type { MappedStates } from '../helpers/custom-state';
-import type { ImportSymbol } from './st-import';
-import type { ClassSymbol } from './css-class';
-import type { ElementSymbol } from './css-type';
-import type { SelectorNode } from '@tokey/css-selector-parser';
-
-export interface StylableDirectives {
-    '-st-root'?: boolean;
-    '-st-states'?: MappedStates;
-    '-st-extends'?: ImportSymbol | ClassSymbol | ElementSymbol;
-    '-st-global'?: SelectorNode[];
-}
