+++ packages/core/test/diagnostics.spec.ts
@@ -5,18 +5,13 @@ import {
     expectTransformDiagnostics,
     findTestLocations,
 } from '@stylable/core-test-kit';
-import {
-    processorDiagnostics,
-    transformerDiagnostics,
-    nativePseudoElements,
-} from '@stylable/core/dist/index-internal';
+import { transformerDiagnostics, nativePseudoElements } from '@stylable/core/dist/index-internal';
 import { CSSClass, CSSType } from '@stylable/core/dist/features';
 import { generalDiagnostics } from '@stylable/core/dist/features/diagnostics';
 
 const cssTypeDiagnostics = diagnosticBankReportToStrings(CSSType.diagnostics);
 const cssClassDiagnostics = diagnosticBankReportToStrings(CSSClass.diagnostics);
 const transformerStringDiagnostics = diagnosticBankReportToStrings(transformerDiagnostics);
-const processorStringDiagnostics = diagnosticBankReportToStrings(processorDiagnostics);
 const generalStringDiagnostics = diagnosticBankReportToStrings(generalDiagnostics);
 
 describe('findTestLocations', () => {
@@ -333,10 +328,7 @@ describe('diagnostics: warnings and errors', () => {
                 `,
                     [
                         {
-                            message: processorStringDiagnostics.OVERRIDE_TYPED_RULE(
-                                `-st-extends`,
-                                'root'
-                            ),
+                            message: cssClassDiagnostics.OVERRIDE_TYPED_RULE(`-st-extends`, 'root'),
                             file: 'main.css',
                         },
                     ]
